move parse_rdb back into utils
