+++ wdfn-server/waterdata/utils.py
@@ -46,6 +46,34 @@ def construct_url(netloc, path, parameters=()):
     return urljoin(netloc, '{0}?{1}'.format(path, encoded_parameters))
 
 
+def parse_rdb(rdb_iter_lines):
+    """
+    Parse records in an RDB file into dictionaries.
+
+    :param iterator rdb_iter_lines: iterator containing lines from an RDB file
+    :rtype: Iterator
+
+    """
+    found_header = False
+    headers = []
+    while not found_header:
+        try:
+            line = next(rdb_iter_lines)
+        except StopIteration:
+            raise Exception('RDB column headers not found.')
+        else:
+            if line and line[0] != '#':
+                headers = line.split('\t')
+                found_header = True
+    # skip the next line in the RDB file
+    next(rdb_iter_lines)
+    for record in rdb_iter_lines:
+        # Ignore empty lines
+        if not record.strip():
+            continue
+        record_values = record.split('\t')
+        yield dict(zip(headers, record_values))
+
 def defined_when(condition, fallback):
     """
     Decorator that fallsback to a specified function if `condition` is False.
